requirements.txt
alembic==0.8.8
bcrypt==3.1.0
bleach==1.4.3
blinker==1.4
certifi==2018.8.24
cffi==1.7.0
chardet==3.0.4
click==6.6
cov-core==1.15.0
coverage==4.2
Django==2.0
djangorestframework==3.8.2
duvet==0.1.2
Flask==0.11.1
Flask-Bcrypt==0.7.1
Flask-HTTPAuth==3.2.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.0.0
Flask-PageDown==0.2.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Uploads==0.2.1
Flask-WTF==0.12
freeze==1.0.10
gunicorn==19.6.0
html5lib==0.9999999
httplib2==0.10.3
idna==2.7
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
Markdown==2.6.7
MarkupSafe==0.23
mysqlclient==1.3.13
nose2==0.6.4
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycparser==2.14
Pygments==2.1.3
PySocks==1.6.6
python-editor==1.0.1
pytz==2016.10
requests==2.11.1
requests-toolbelt==0.7.0
six==1.10.0
SQLAlchemy==1.0.13
tkreadonly==0.6.0
twilio==5.7.0
urllib3==1.23
webencodings==0.5.1
Werkzeug==0.11.10
WTForms==2.1
model_mommy